backend/requirements.txt
ultralytics
opencv-python
fastapi
uvicorn
python-socketio
openvino-dev
